# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Todo API Documentation",
        "version": "1.0.0"
    },
    "basePath": "/api/",
    "paths": {
        "/todos": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "Get a listing of the Todos.",
                "description": "Get all Todos",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "state",
                        "in": "query",
                        "description": "search by state (done, undone)",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "done",
                            "undone"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Todo"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Todo"
                ],
                "summary": "Create a new todo",
                "description": "Store Todo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Todo that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/newTodo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Todo"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "Display the specified Todo",
                "description": "Get Todo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Todo",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Todo"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Todo"
                ],
                "summary": "Update the status of the todo",
                "description": "Update Todo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Todo",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "update the state of the todo",
                        "required": false,
                        "schema": {
                            "properties": {
                                "state": {
                                    "description": "state of todo (done or undone)",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Todo"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todo"
                ],
                "summary": "Remove the specified Todo by id",
                "description": "Delete Todo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Todo",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Todo": {
            "required": [
                "name",
                "activity",
                "state",
                "dateline"
            ],
            "properties": {
                "todos": {
                    "description": "todos",
                    "properties": {
                        "id": {
                            "description": "id",
                            "type": "integer",
                            "format": "int32"
                        },
                        "name": {
                            "description": "name",
                            "type": "string"
                        },
                        "activity": {
                            "description": "activity",
                            "type": "string"
                        },
                        "state": {
                            "description": "state",
                            "type": "string"
                        },
                        "dateline": {
                            "description": "dateline",
                            "type": "string",
                            "format": "date"
                        },
                        "created_at": {
                            "description": "created_at",
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    "type": "object"
                },
                "productivity_score": {
                    "description": "productivity_score",
                    "type": "string",
                    "format": "0.2"
                }
            }
        },
        "newTodo": {
            "required": [
                "name",
                "activity",
                "dateline"
            ],
            "properties": {
                "name": {
                    "description": "name",
                    "type": "string"
                },
                "activity": {
                    "description": "activity",
                    "type": "string"
                },
                "dateline": {
                    "description": "dateline",
                    "type": "string",
                    "format": "date"
                }
            }
        }
    }
}